// src/app/position/model/position.model.ts
export class PositionModel {
    id: number;
    type_launch: string;
    description: string;
    date_last_purchase: string;
    total_debit_week: number;
    total_credit_week: number;
    total_apportionment: string;
    synchronized: string;
    observation: string;
    reminder: string;
    author: number;

    // ID: number;
    // nameActive: string;
    // percentActive: string;
    // colorActive: string;
}